@@ -33,8 +33,9 @@ from ..utils.meas_info import _HUMAN_UNITS, _set_channel_units
 from ._filters import StreamFilter, create_filter, ensure_sos_iir_params
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
     from datetime import datetime
-    from typing import Any, Callable, Optional, Union
+    from typing import Any
 
     from mne import Info
     from mne.channels import DigMontage
@@ -110,10 +111,8 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def add_reference_channels(
         self,
-        ref_channels: Union[str, list[str], tuple[str]],
-        ref_units: Optional[
-            Union[str, int, list[Union[str, int]], tuple[Union[str, int]]]
-        ] = None,
+        ref_channels: str | list[str] | tuple[str, ...],
+        ref_units: str | int | list[str | int] | tuple[str | int, ...] | None = None,
     ) -> BaseStream:
         """Add EEG reference channels to data that consists of all zeros.
 
@@ -156,7 +155,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
         # error checking and conversion of the arguments to valid values
         if isinstance(ref_channels, str):
             ref_channels = [ref_channels]
-        if isinstance(ref_units, (str, int)):
+        if isinstance(ref_units, (str | int)):
             ref_units = [ref_units]
         elif ref_units is None:
             ref_units = [0] * len(ref_channels)
@@ -238,10 +237,10 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def anonymize(
         self,
-        daysback: Optional[int] = None,
+        daysback: int | None = None,
         keep_his: bool = False,
         *,
-        verbose: Optional[Union[bool, str, int]] = None,
+        verbose: bool | str | int | None = None,
     ) -> BaseStream:
         """Anonymize the measurement information in-place.
 
@@ -340,7 +339,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
         # This method needs to close any inlet/network object and need to end with
         # self._reset_variables().
 
-    def del_filter(self, idx: Union[int, list[int], tuple[int], str] = "all") -> None:
+    def del_filter(self, idx: int | list[int] | tuple[int, ...] | str = "all") -> None:
         """Remove a filter from the list of applied filters.
 
         Parameters
@@ -373,7 +372,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
             )
         elif idx == "all":
             idx = np.arange(len(self._filters), dtype=np.uint8)
-        elif isinstance(idx, (tuple, list)):
+        elif isinstance(idx, (tuple | list)):
             for elt in idx:
                 check_type(elt, ("int-like",), "idx")
             idx = np.array(idx, dtype=np.uint8)
@@ -419,7 +418,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
             for k in idx[::-1]:
                 del self._filters[k]
 
-    def drop_channels(self, ch_names: Union[str, list[str], tuple[str]]) -> BaseStream:
+    def drop_channels(self, ch_names: str | list[str] | tuple[str, ...]) -> BaseStream:
         """Drop channel(s).
 
         Parameters
@@ -456,12 +455,12 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def filter(
         self,
-        l_freq: Optional[float],
-        h_freq: Optional[float],
-        picks: Optional[Union[str, list[str], int, list[int], ScalarIntArray]] = None,
-        iir_params: Optional[dict[str, Any]] = None,
+        l_freq: float | None,
+        h_freq: float | None,
+        picks: str | list[str] | int | list[int] | ScalarIntArray | None = None,
+        iir_params: dict[str, Any] | None = None,
         *,
-        verbose: Optional[Union[bool, str, int]] = None,
+        verbose: bool | str | int | None = None,
     ) -> BaseStream:  # noqa: A003
         """Filter the stream with an IIR causal filter.
 
@@ -517,7 +516,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @copy_doc(ContainsMixin.get_channel_types)
     def get_channel_types(
         self,
-        picks: Optional[Union[str, list[str], int, list[int], ScalarIntArray]] = None,
+        picks: str | list[str] | int | list[int] | ScalarIntArray | None = None,
         unique=False,
         only_data_chs=False,
     ) -> list[str]:
@@ -529,7 +528,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def get_channel_units(
         self,
-        picks: Optional[Union[str, list[str], int, list[int], ScalarIntArray]] = None,
+        picks: str | list[str] | int | list[int] | ScalarIntArray | None = None,
         only_data_chs: bool = False,
     ) -> list[tuple[int, int]]:
         """Get a list of channel unit for each channel.
@@ -562,9 +561,9 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def get_data(
         self,
-        winsize: Optional[float] = None,
-        picks: Optional[Union[str, list[str], int, list[int], ScalarIntArray]] = None,
-        exclude: Union[str, list[str], tuple[str]] = "bads",
+        winsize: float | None = None,
+        picks: str | list[str] | int | list[int] | ScalarIntArray | None = None,
+        exclude: str | list[str] | tuple[str, ...] = "bads",
     ) -> tuple[ScalarArray, NDArray[np.float64]]:
         """Retrieve the latest data from the buffer.
 
@@ -630,7 +629,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
             raise  # pragma: no cover
 
     @copy_doc(SetChannelsMixin.get_montage)
-    def get_montage(self) -> Optional[DigMontage]:
+    def get_montage(self) -> DigMontage | None:
         self._check_connected("get_montage()")
         return super().get_montage()
 
@@ -639,12 +638,12 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     def notch_filter(
         self,
         freqs: float,
-        picks: Optional[Union[str, list[str], int, list[int], ScalarIntArray]] = None,
-        notch_widths: Optional[float] = None,
+        picks: str | list[str] | int | list[int] | ScalarIntArray | None = None,
+        notch_widths: float | None = None,
         trans_bandwidth=1,
-        iir_params: Optional[dict[str, Any]] = None,
+        iir_params: dict[str, Any] | None = None,
         *,
-        verbose: Optional[Union[bool, str, int]] = None,
+        verbose: bool | str | int | None = None,
     ) -> BaseStream:
         """Filter the stream with an IIR causal notch filter.
 
@@ -734,8 +733,8 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def pick(
         self,
-        picks: Optional[Union[str, list[str], int, list[int], ScalarIntArray]] = None,
-        exclude: Union[str, list[str], int, list[int], ScalarIntArray] = (),
+        picks: str | list[str] | int | list[int] | ScalarIntArray | None = None,
+        exclude: str | list[str] | int | list[int] | ScalarIntArray = (),
     ) -> BaseStream:
         """Pick a subset of channels.
 
@@ -777,10 +776,10 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def rename_channels(
         self,
-        mapping: Union[dict[str, str], Callable],
+        mapping: dict[str, str] | Callable,
         allow_duplicates: bool = False,
         *,
-        verbose: Optional[Union[bool, str, int]] = None,
+        verbose: bool | str | int | None = None,
     ) -> BaseStream:
         """Rename channels.
 
@@ -829,7 +828,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
         mapping: dict[str, str],
         *,
         on_unit_change: str = "warn",
-        verbose: Optional[Union[bool, str, int]] = None,
+        verbose: bool | str | int | None = None,
     ) -> BaseStream:
         """Define the sensor type of channels.
 
@@ -864,7 +863,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
         )
         return self
 
-    def set_channel_units(self, mapping: dict[str, Union[str, int]]) -> BaseStream:
+    def set_channel_units(self, mapping: dict[str, str | int]) -> BaseStream:
         """Define the channel unit multiplication factor.
 
         The unit itself is defined by the sensor type. Use
@@ -896,8 +895,8 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def set_eeg_reference(
         self,
-        ref_channels: Union[str, list[str], tuple[str]],
-        ch_type: Union[str, list[str], tuple[str]] = "eeg",
+        ref_channels: str | list[str] | tuple[str, ...],
+        ch_type: str | list[str] | tuple[str, ...] = "eeg",
     ) -> BaseStream:
         """Specify which reference to use for EEG-like data.
 
@@ -969,7 +968,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
         return self
 
     def set_meas_date(
-        self, meas_date: Optional[Union[datetime, float, tuple[float]]]
+        self, meas_date: datetime | float | tuple[float, float] | None
     ) -> BaseStream:
         """Set the measurement start date.
 
@@ -1001,12 +1000,12 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     @fill_doc
     def set_montage(
         self,
-        montage: Optional[Union[str, DigMontage]],
+        montage: str | DigMontage | None,
         match_case: bool = True,
-        match_alias: Union[bool, dict[str, str]] = False,
+        match_alias: bool | dict[str, str] = False,
         on_missing: str = "raise",
         *,
-        verbose: Optional[Union[bool, str, int]] = None,
+        verbose: bool | str | int | None = None,
     ) -> BaseStream:
         """Set %(montage_types)s channel positions and digitization points.
 
@@ -1158,7 +1157,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
 
     # ----------------------------------------------------------------------------------
     @property
-    def compensation_grade(self) -> Optional[int]:
+    def compensation_grade(self) -> int | None:
         """The current gradient compensation grade.
 
         :type: :class:`int` | None
@@ -1198,7 +1197,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
             return True
 
     @property
-    def dtype(self) -> Optional[DTypeLike]:
+    def dtype(self) -> DTypeLike | None:
         """Channel format of the stream."""
         return getattr(self._buffer, "dtype", None)
 
