@@ -10,7 +10,6 @@ from numpy.testing import assert_allclose, assert_array_equal
 
 if TYPE_CHECKING:
     from pathlib import Path
-    from typing import Union
 
     from mne import Info
     from mne.io import BaseRaw
@@ -18,7 +17,7 @@ if TYPE_CHECKING:
     from .._typing import ScalarArray
 
 
-def sha256sum(fname: Union[str, Path]) -> str:
+def sha256sum(fname: str | Path) -> str:
     """Efficiently hash a file."""
     h = hashlib.sha256()
     b = bytearray(128 * 1024)
@@ -112,7 +111,7 @@ def compare_infos(info1: Info, info2: Info) -> None:
     assert len(info1["projs"]) == len(info2["projs"])
     projs1 = sorted(info1["projs"], key=lambda x: x["desc"])
     projs2 = sorted(info2["projs"], key=lambda x: x["desc"])
-    for proj1, proj2 in zip(projs1, projs2):
+    for proj1, proj2 in zip(projs1, projs2, strict=True):
         assert proj1["desc"] == proj2["desc"]
         assert proj1["kind"] == proj2["kind"]
         assert proj1["data"]["nrow"] == proj2["data"]["nrow"]
@@ -125,7 +124,7 @@ def compare_infos(info1: Info, info2: Info) -> None:
         assert len(info1["dig"]) == len(info2["dig"])
         digs1 = sorted(info1["dig"], key=lambda x: (x["kind"], x["ident"]))
         digs2 = sorted(info2["dig"], key=lambda x: (x["kind"], x["ident"]))
-        for dig1, dig2 in zip(digs1, digs2):
+        for dig1, dig2 in zip(digs1, digs2, strict=True):
             assert dig1["kind"] == dig2["kind"]
             assert dig1["ident"] == dig2["ident"]
             assert dig1["coord_frame"] == dig2["coord_frame"]
