@@ -1,8 +1,9 @@
 import platform
 import sys
+from collections.abc import Callable
 from functools import lru_cache, partial
 from importlib.metadata import requires, version
-from typing import IO, Callable, Optional
+from typing import IO
 
 import psutil
 from packaging.requirements import Requirement
@@ -11,7 +12,7 @@ from ._checks import check_type
 from .logs import _use_log_level
 
 
-def sys_info(fid: Optional[IO] = None, developer: bool = False):
+def sys_info(fid: IO | None = None, developer: bool = False):
     """Print the system information for debugging.
 
     Parameters
@@ -160,7 +161,7 @@ def _list_dependencies_info(
 
 
 @lru_cache(maxsize=1)
-def _get_gpu_info() -> tuple[Optional[str], Optional[str]]:
+def _get_gpu_info() -> tuple[str | None, str | None]:
     """Get the GPU information."""
     try:
         from pyvista import GPUInfo
