@@ -1,7 +1,7 @@
 import logging
+from collections.abc import Callable
 from logging import Logger
 from pathlib import Path
-from typing import Callable
 
 from _typeshed import Incomplete
 
